Relocate non-UI search core to metabase/utils/search (lib tier)
Moves the search filter types, SearchFilterKeys constants, analytics
helpers, and URL/location utils out of metabase/search into the lib-tier
metabase/utils/search so non-feature modules (api, palette, plugins,
entity pickers) can keep importing them once search becomes a feature
module. Pure relocation + import-path rewrites; no behavior change.
