+++ frontend/src/metabase/api/search.ts
@@ -1,4 +1,4 @@
-import { isTrackedSearchContext } from "metabase/search/analytics";
+import { isTrackedSearchContext } from "metabase/utils/search/analytics";
 import type { SearchRequest, SearchResponse } from "metabase-types/api";
 
 import { trackSearchRequest } from "./analytics";
