+++ frontend/src/metabase/nav/components/search/SearchResults/SearchResults.tsx
@@ -14,13 +14,13 @@ import {
 } from "metabase/nav/components/search/SearchResults/SearchResults.styled";
 import { useDispatch } from "metabase/redux";
 import { SearchResult } from "metabase/search/components/SearchResult/SearchResult";
-import type { SearchFilters } from "metabase/search/types";
 import { Loader } from "metabase/ui";
 import { modelToUrl } from "metabase/urls";
 import {
   DEFAULT_SEARCH_LIMIT,
   SEARCH_DEBOUNCE_DURATION,
 } from "metabase/utils/constants";
+import type { SearchFilters } from "metabase/utils/search/types";
 import type {
   CollectionItem,
   SearchContext,
