+++ frontend/src/metabase/palette/hooks/useCommandPalette.tsx
@@ -11,7 +11,6 @@ import type { OmniPickerItem } from "metabase/common/components/Pickers";
 import { useSetting } from "metabase/common/hooks";
 import { useGetIcon } from "metabase/hooks/use-icon";
 import { useSelector } from "metabase/redux";
-import { trackSearchClick } from "metabase/search/analytics";
 import {
   getDocsSearchUrl,
   getDocsUrl,
@@ -24,6 +23,7 @@ import * as Urls from "metabase/urls";
 import { modelToUrl } from "metabase/urls";
 import { SEARCH_DEBOUNCE_DURATION } from "metabase/utils/constants";
 import { getName } from "metabase/utils/name";
+import { trackSearchClick } from "metabase/utils/search/analytics";
 import {
   type RecentCollectionItem,
   type RecentItem,
