+++ frontend/src/metabase/search/components/SearchSidebar/tests/setup.tsx
@@ -4,7 +4,7 @@ import { mockSettings } from "__support__/settings";
 import { renderWithProviders } from "__support__/ui";
 import { createMockState } from "metabase/redux/store/mocks";
 import { SearchSidebar } from "metabase/search/components/SearchSidebar";
-import type { URLSearchFilterQueryParams } from "metabase/search/types";
+import type { URLSearchFilterQueryParams } from "metabase/utils/search/types";
 import type { TokenFeatures } from "metabase-types/api";
 import {
   createMockDatabase,
