+++ frontend/src/metabase/search/components/filters/LastEditedAtFilter.tsx
@@ -2,8 +2,8 @@ import { t } from "ttag";
 
 import { SearchFilterDateDisplay } from "metabase/search/components/SearchFilterDateDisplay";
 import { SearchFilterDatePicker } from "metabase/search/components/SearchFilterDatePicker";
-import type { SearchFilterDropdown } from "metabase/search/types";
 import { Box } from "metabase/ui";
+import type { SearchFilterDropdown } from "metabase/utils/search/types";
 
 export const LastEditedAtFilter: SearchFilterDropdown<"last_edited_at"> = {
   iconName: "calendar",
