+++ frontend/src/metabase/search/components/filters/LastEditedByFilter.tsx
@@ -3,8 +3,8 @@ import { t } from "ttag";
 import { SearchUserPicker } from "metabase/search/components/SearchUserPicker/SearchUserPicker";
 import { SearchUserPickerContainer } from "metabase/search/components/SearchUserPicker/SearchUserPicker.styled";
 import { UserNameDisplay } from "metabase/search/components/UserNameDisplay/UserNameDisplay";
-import type { SearchFilterDropdown } from "metabase/search/types";
-import { parseUserIdArray, stringifyUserIdArray } from "metabase/search/utils";
+import { parseUserIdArray, stringifyUserIdArray } from "metabase/utils/search";
+import type { SearchFilterDropdown } from "metabase/utils/search/types";
 
 export const LastEditedByFilter: SearchFilterDropdown<"last_edited_by"> = {
   iconName: "person",
