+++ frontend/src/metabase/search/components/filters/NativeQueryFilter/NativeQueryFilter.tsx
@@ -1,5 +1,5 @@
 import { NativeQueryLabel } from "metabase/search/components/filters/NativeQueryFilter/NativeQueryLabel";
-import type { SearchFilterToggle } from "metabase/search/types";
+import type { SearchFilterToggle } from "metabase/utils/search/types";
 
 export const NativeQueryFilter: SearchFilterToggle = {
   label: NativeQueryLabel,
