+++ frontend/src/metabase/search/components/filters/TypeFilter/TypeFilterContent.tsx
@@ -3,9 +3,9 @@ import { useState } from "react";
 import { useSearchQuery } from "metabase/api";
 import { getTranslatedEntityName } from "metabase/common/utils/model-names";
 import { SearchFilterPopoverWrapper } from "metabase/search/components/SearchFilterPopoverWrapper";
-import { enabledSearchTypes } from "metabase/search/constants";
-import type { SearchFilterDropdown } from "metabase/search/types";
 import { Checkbox, Stack } from "metabase/ui";
+import { enabledSearchTypes } from "metabase/utils/search/constants";
+import type { SearchFilterDropdown } from "metabase/utils/search/types";
 import type { EnabledSearchModel } from "metabase-types/api";
 
 const EMPTY_SEARCH_QUERY = {
