+++ frontend/src/metabase/search/components/filters/TypeFilter/TypeFilterDisplay.tsx
@@ -2,8 +2,8 @@ import { t } from "ttag";
 
 import { getTranslatedEntityName } from "metabase/common/utils/model-names";
 import { TypeFilter } from "metabase/search/components/filters/TypeFilter";
-import type { SearchFilterDropdown } from "metabase/search/types";
 import { Text } from "metabase/ui";
+import type { SearchFilterDropdown } from "metabase/utils/search/types";
 
 export const TypeFilterDisplay: SearchFilterDropdown<"type">["DisplayComponent"] =
   ({ value }) => {
