+++ frontend/src/metabase/search/containers/SearchApp.unit.spec.tsx
@@ -15,7 +15,7 @@ import {
   within,
 } from "__support__/ui";
 import { SearchApp } from "metabase/search/containers/SearchApp";
-import type { SearchFilters } from "metabase/search/types";
+import type { SearchFilters } from "metabase/utils/search/types";
 import { checkNotNull } from "metabase/utils/types";
 import type { EnabledSearchModel, SearchResult } from "metabase-types/api";
 import {
