+++ frontend/src/metabase/utils/search/constants.ts
