+++ frontend/src/metabase/utils/search/enabled-search-type/enabled-search-type.ts
@@ -1,4 +1,4 @@
-import { enabledSearchTypes } from "metabase/search/constants";
+import { enabledSearchTypes } from "metabase/utils/search/constants";
 import type { EnabledSearchModel } from "metabase-types/api";
 
 export function isEnabledSearchModelType(
