+++ frontend/src/metabase/utils/search/enabled-search-type/index.ts
