+++ frontend/src/metabase/utils/search/index.ts
@@ -1,3 +1,7 @@
+export * from "./term";
+export * from "./analytics";
+export * from "./constants";
+export * from "./types";
 export * from "./search-location";
 export * from "./user-search-params";
 export * from "./enabled-search-type";
