+++ frontend/src/metabase/utils/search/search-location/index.ts
