+++ frontend/src/metabase/utils/search/search-location/search-location.ts
@@ -1,10 +1,10 @@
 import _ from "underscore";
 
-import { SearchFilterKeys } from "metabase/search/constants";
+import { SearchFilterKeys } from "metabase/utils/search/constants";
 import type {
   SearchAwareLocation,
   URLSearchFilterQueryParams,
-} from "metabase/search/types";
+} from "metabase/utils/search/types";
 
 export function isSearchPageLocation(location?: SearchAwareLocation): boolean {
   return location ? /^\/?search$/.test(location.pathname) : false;
