+++ frontend/src/metabase/utils/search/search-location/search-location.unit.spec.ts
@@ -1,10 +1,10 @@
-import { SearchFilterKeys } from "metabase/search/constants";
-import type { SearchAwareLocation } from "metabase/search/types";
 import {
   getFiltersFromLocation,
   getSearchTextFromLocation,
   isSearchPageLocation,
-} from "metabase/search/utils";
+} from "metabase/utils/search";
+import { SearchFilterKeys } from "metabase/utils/search/constants";
+import type { SearchAwareLocation } from "metabase/utils/search/types";
 
 describe("isSearchPageLocation", () => {
   it("should return true for a search page location", () => {
