+++ frontend/src/metabase/utils/search/term.ts
