+++ frontend/src/metabase/utils/search/user-search-params/user-search-params.ts
@@ -1,4 +1,4 @@
-import type { SearchQueryParamValue } from "metabase/search/types";
+import type { SearchQueryParamValue } from "metabase/utils/search/types";
 import { isNotNull } from "metabase/utils/types";
 import type { UserId } from "metabase-types/api";
